Moves state_array into adt
